# mw-general-bff/apiAutogenerated/mw-general-bff/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  title: Masters way general-bff API
  version: "1.0"
servers:
- url: /general
paths:
...
  /wayCollections:
    post:
      operationId: create-wayCollection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreateWayCollectionPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayCollectionPopulatedResponse'
          description: OK
      summary: Create a new wayCollection
      tags:
      - wayCollection
      x-codegen-request-body-name: request
  /wayCollections/{wayCollectionId}:
    delete:
      operationId: delete-wayCollection
      parameters:
      - description: wayCollection ID
        in: path
        name: wayCollectionId
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete wayCollection by UUID
      tags:
      - wayCollection
    patch:
      operationId: update-wayCollection
      parameters:
      - description: wayCollection ID
        in: path
        name: wayCollectionId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.UpdateWayCollectionPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayCollectionPlainResponse'
          description: OK
      summary: Update wayCollection by UUID
      tags:
      - wayCollection
      x-codegen-request-body-name: request
...
  /ways:
    get:
      description: Get ways with pagination
      operationId: get-all-ways
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        schema:
          type: integer
      - description: Number of items per page
        in: query
        name: limit
        schema:
          type: integer
      - description: Min day reports amount
        in: query
        name: minDayReportsAmount
        schema:
          type: integer
      - description: Way name
        in: query
        name: wayName
        schema:
          type: string
      - description: "Ways type: all | completed | inProgress | abandoned"
        in: query
        name: status
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.GetAllWaysResponse'
          description: OK
      summary: Get all ways
      tags:
      - way
    post:
      operationId: create-way
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreateWayPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayPlainResponse'
          description: OK
      summary: Create a new way
      tags:
      - way
      x-codegen-request-body-name: request
  /ways/createFromTraining:
    post:
      description: Create a new way based on training
      operationId: create-way-from-training
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreateWayFromTrainingPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreateWayFromTrainingResponse'
          description: OK
      summary: Create a new way based on training
      tags:
      - way
      x-codegen-request-body-name: request
  /ways/{wayId}:
    delete:
      operationId: delete-way
      parameters:
      - description: way ID
        in: path
        name: wayId
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete way by UUID
      tags:
      - way
    get:
      operationId: get-way-by-uuid
      parameters:
      - description: way ID
        in: path
        name: wayId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayPopulatedResponse'
          description: OK
      summary: Get way by UUID
      tags:
      - way
    patch:
      operationId: update-way
      parameters:
      - description: way ID
        in: path
        name: wayId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.UpdateWayPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayPlainResponse'
          description: OK
      summary: Update way by UUID
      tags:
      - way
      x-codegen-request-body-name: request
  /ways/{wayId}/statistics:
    get:
      operationId: get-way-statistics-by-uuid
      parameters:
      - description: way ID
        in: path
        name: wayId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayStatisticsTriplePeriod'
          description: OK
      summary: Get way statistics by UUID
      tags:
      - way
components:
  schemas:
...
    mw-general-bff_internal_schemas.CreateWayCollectionPayload:
      properties:
        name:
          type: string
        ownerUuid:
          type: string
      required:
      - name
      - ownerUuid
      type: object
...
    mw-general-bff_internal_schemas.CreateWayFromTrainingPayload:
      properties:
        trainingId:
          type: string
      required:
      - trainingId
      type: object
    mw-general-bff_internal_schemas.CreateWayFromTrainingResponse:
      example:
        wayUuid: wayUuid
      properties:
        wayUuid:
          type: string
      required:
      - wayUuid
      type: object
    mw-general-bff_internal_schemas.CreateWayPayload:
      properties:
        copiedFromWayId:
          nullable: true
          type: string
        estimationTime:
          type: integer
        goalDescription:
          type: string
        isCompleted:
          type: boolean
        isPrivate:
          type: boolean
        name:
          type: string
        ownerId:
          type: string
        projectId:
          nullable: true
          type: string
      required:
      - copiedFromWayId
      - estimationTime
      - goalDescription
      - isCompleted
      - isPrivate
      - name
      - ownerId
      - projectId
      type: object
...
    mw-general-bff_internal_schemas.GetAllWaysResponse:
      example:
        ways:
        - owner:
            createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          goalDescription: goalDescription
          copiedFromWayUuid: copiedFromWayUuid
          childrenUuids:
          - childrenUuids
          - childrenUuids
          estimationTime: 6
          isPrivate: true
          uuid: uuid
          dayReportsAmount: 0
          projectUuid: projectUuid
          metricsDone: 5
          createdAt: createdAt
          wayTags:
          - name: name
            uuid: uuid
          - name: name
            uuid: uuid
          favoriteForUsers: 1
          metricsTotal: 5
          mentors:
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          name: name
          isCompleted: true
          updatedAt: updatedAt
        - owner:
            createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          goalDescription: goalDescription
          copiedFromWayUuid: copiedFromWayUuid
          childrenUuids:
          - childrenUuids
          - childrenUuids
          estimationTime: 6
          isPrivate: true
          uuid: uuid
          dayReportsAmount: 0
          projectUuid: projectUuid
          metricsDone: 5
          createdAt: createdAt
          wayTags:
          - name: name
            uuid: uuid
          - name: name
            uuid: uuid
          favoriteForUsers: 1
          metricsTotal: 5
          mentors:
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          name: name
          isCompleted: true
          updatedAt: updatedAt
        size: 0
      properties:
        size:
          type: integer
        ways:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayPlainResponse'
          type: array
      required:
      - size
      - ways
      type: object
...
    mw-general-bff_internal_schemas.UpdateWayCollectionPayload:
      properties:
        name:
          type: string
      type: object
    mw-general-bff_internal_schemas.UpdateWayPayload:
      properties:
        estimationTime:
          type: integer
        goalDescription:
          type: string
        isCompleted:
          type: boolean
        isPrivate:
          type: boolean
        name:
          type: string
      type: object
...
    mw-general-bff_internal_schemas.WayCollectionPlainResponse:
      example:
        name: name
        uuid: uuid
      properties:
        name:
          type: string
        uuid:
          type: string
      required:
      - name
      - uuid
      type: object
    mw-general-bff_internal_schemas.WayCollectionPopulatedResponse:
      example:
        createdAt: createdAt
        ownerUuid: ownerUuid
        ways:
        - owner:
            createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          goalDescription: goalDescription
          copiedFromWayUuid: copiedFromWayUuid
          childrenUuids:
          - childrenUuids
          - childrenUuids
          estimationTime: 6
          isPrivate: true
          uuid: uuid
          dayReportsAmount: 0
          projectUuid: projectUuid
          metricsDone: 5
          createdAt: createdAt
          wayTags:
          - name: name
            uuid: uuid
          - name: name
            uuid: uuid
          favoriteForUsers: 1
          metricsTotal: 5
          mentors:
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          name: name
          isCompleted: true
          updatedAt: updatedAt
        - owner:
            createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          goalDescription: goalDescription
          copiedFromWayUuid: copiedFromWayUuid
          childrenUuids:
          - childrenUuids
          - childrenUuids
          estimationTime: 6
          isPrivate: true
          uuid: uuid
          dayReportsAmount: 0
          projectUuid: projectUuid
          metricsDone: 5
          createdAt: createdAt
          wayTags:
          - name: name
            uuid: uuid
          - name: name
            uuid: uuid
          favoriteForUsers: 1
          metricsTotal: 5
          mentors:
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          - createdAt: createdAt
            imageUrl: imageUrl
            name: name
            description: description
            isMentor: true
            uuid: uuid
            email: email
          name: name
          isCompleted: true
          updatedAt: updatedAt
        name: name
        type: type
        uuid: uuid
        updatedAt: updatedAt
      properties:
        createdAt:
          type: string
        name:
          type: string
        ownerUuid:
          type: string
        type:
          description: should be removed after separation custom collections and default
            pseudocollections
          type: string
        updatedAt:
          type: string
        uuid:
          type: string
        ways:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayPlainResponse'
          type: array
      required:
      - createdAt
      - name
      - ownerUuid
      - type
      - updatedAt
      - uuid
      - ways
      type: object
    mw-general-bff_internal_schemas.WayPlainResponse:
      example:
        owner:
          createdAt: createdAt
          imageUrl: imageUrl
          name: name
          description: description
          isMentor: true
          uuid: uuid
          email: email
        goalDescription: goalDescription
        copiedFromWayUuid: copiedFromWayUuid
        childrenUuids:
        - childrenUuids
        - childrenUuids
        estimationTime: 6
        isPrivate: true
        uuid: uuid
        dayReportsAmount: 0
        projectUuid: projectUuid
        metricsDone: 5
        createdAt: createdAt
        wayTags:
        - name: name
          uuid: uuid
        - name: name
          uuid: uuid
        favoriteForUsers: 1
        metricsTotal: 5
        mentors:
        - createdAt: createdAt
          imageUrl: imageUrl
          name: name
          description: description
          isMentor: true
          uuid: uuid
          email: email
        - createdAt: createdAt
          imageUrl: imageUrl
          name: name
          description: description
          isMentor: true
          uuid: uuid
          email: email
        name: name
        isCompleted: true
        updatedAt: updatedAt
      properties:
        childrenUuids:
          items:
            type: string
          type: array
        copiedFromWayUuid:
          nullable: true
          type: string
        createdAt:
          type: string
        dayReportsAmount:
          type: integer
        estimationTime:
          type: integer
        favoriteForUsers:
          type: integer
        goalDescription:
          type: string
        isCompleted:
          type: boolean
        isPrivate:
          type: boolean
        mentors:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.UserPlainResponse'
          type: array
        metricsDone:
          type: integer
        metricsTotal:
          type: integer
        name:
          type: string
        owner:
          $ref: '#/components/schemas/mw-general-bff_internal_schemas.UserPlainResponse'
        projectUuid:
          nullable: true
          type: string
        updatedAt:
          type: string
        uuid:
          type: string
        wayTags:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayTagResponse'
          type: array
      required:
      - childrenUuids
      - copiedFromWayUuid
      - createdAt
      - dayReportsAmount
      - estimationTime
      - favoriteForUsers
      - goalDescription
      - isCompleted
      - isPrivate
      - mentors
      - metricsDone
      - metricsTotal
      - name
      - owner
      - projectUuid
      - updatedAt
      - uuid
      - wayTags
      type: object
...
    mw-general-bff_internal_schemas.WayStatisticsTriplePeriod:
      example:
        totalTime:
          labelStatistics:
            labels:
            - jobsAmount: 0
              jobsAmountPercentage: 6
              label:
                color: color
                name: name
                description: description
                uuid: uuid
              time: 1
              timePercentage: 5
            - jobsAmount: 0
              jobsAmountPercentage: 6
              label:
                color: color
                name: name
                description: description
                uuid: uuid
              time: 1
              timePercentage: 5
          timeSpentByDayChart:
          - date: date
            value: 4
          - date: date
            value: 4
          overallInformation:
            averageTimePerWorkingDay: 7
            totalTime: 2
            averageTimePerCalendarDay: 2
            totalReports: 3
            finishedJobs: 9
            averageJobTime: 5
        lastMonth:
          labelStatistics:
            labels:
            - jobsAmount: 0
              jobsAmountPercentage: 6
              label:
                color: color
                name: name
                description: description
                uuid: uuid
              time: 1
              timePercentage: 5
            - jobsAmount: 0
              jobsAmountPercentage: 6
              label:
                color: color
                name: name
                description: description
                uuid: uuid
              time: 1
              timePercentage: 5
          timeSpentByDayChart:
          - date: date
            value: 4
          - date: date
            value: 4
          overallInformation:
            averageTimePerWorkingDay: 7
            totalTime: 2
            averageTimePerCalendarDay: 2
            totalReports: 3
            finishedJobs: 9
            averageJobTime: 5
        lastWeek:
          labelStatistics:
            labels:
            - jobsAmount: 0
              jobsAmountPercentage: 6
              label:
                color: color
                name: name
                description: description
                uuid: uuid
              time: 1
              timePercentage: 5
            - jobsAmount: 0
              jobsAmountPercentage: 6
              label:
                color: color
                name: name
                description: description
                uuid: uuid
              time: 1
              timePercentage: 5
          timeSpentByDayChart:
          - date: date
            value: 4
          - date: date
            value: 4
          overallInformation:
            averageTimePerWorkingDay: 7
            totalTime: 2
            averageTimePerCalendarDay: 2
            totalReports: 3
            finishedJobs: 9
            averageJobTime: 5
      properties:
        lastMonth:
          $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayStatistics'
        lastWeek:
          $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayStatistics'
        totalTime:
          $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayStatistics'
      required:
      - lastMonth
      - lastWeek
      - totalTime
      type: object
    mw-general-bff_internal_schemas.WayTagResponse:
      example:
        name: name
        uuid: uuid
      properties:
        name:
          type: string
        uuid:
          type: string
      required:
      - name
      - uuid
      type: object
...