# mw-general-bff/internal/controllers/wayCollection.controller.go
package controllers

import (
	openapiGeneral "mw-general-bff/apiAutogenerated/general"
	"mw-general-bff/internal/facades"
	"mw-general-bff/internal/schemas"
	"mw-general-bff/pkg/utils"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/samber/lo"
)

// Without next lines swagger does not see openapi models
var _ = &openapiGeneral.MwServerInternalSchemasWayCollectionPopulatedResponse{}

type WayCollectionController struct {
	wayCollectionFacade *facades.WayCollectionFacade
}

func NewWayCollectionController(wayCollectionFacade *facades.WayCollectionFacade) *WayCollectionController {
	return &WayCollectionController{wayCollectionFacade}
}

// Create wayCollectionRoute handler
// @Summary Create a new wayCollection
// @Description
// @Tags wayCollection
// @ID create-wayCollection
// @Accept  json
// @Produce  json
// @Param request body schemas.CreateWayCollectionPayload true "query params"
// @Success 200 {object} schemas.WayCollectionPopulatedResponse
// @Router /wayCollections [post]
func (wc *WayCollectionController) CreateWayCollection(ctx *gin.Context) {
	var payload *schemas.CreateWayCollectionPayload

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	args := &schemas.CreateWayCollectionPayload{
		Name:      payload.Name,
		OwnerUuid: payload.OwnerUuid,
	}
	responseRaw, err := wc.wayCollectionFacade.CreateWayCollection(ctx, args)
	utils.HandleErrorGin(ctx, err)

	response := schemas.WayCollectionPopulatedResponse{
		Uuid:      responseRaw.Uuid,
		Name:      responseRaw.Name,
		Ways:      lo.Map(responseRaw.Ways, ConvertWay),
		CreatedAt: responseRaw.CreatedAt,
		UpdatedAt: responseRaw.UpdatedAt,
		OwnerUuid: responseRaw.OwnerUuid,
		Type:      responseRaw.Type,
	}

	ctx.JSON(http.StatusOK, response)
}

// Update wayCollectionRoute handler
// @Summary Update wayCollection by UUID
// @Description
// @Tags wayCollection
// @ID update-wayCollection
// @Accept  json
// @Produce  json
// @Param request body schemas.UpdateWayCollectionPayload true "query params"
// @Param wayCollectionId path string true "wayCollection ID"
// @Success 200 {object} schemas.WayCollectionPlainResponse
// @Router /wayCollections/{wayCollectionId} [patch]
func (wc *WayCollectionController) UpdateWayCollection(ctx *gin.Context) {
	var payload *schemas.UpdateWayCollectionPayload
	wayCollectionID := ctx.Param("wayCollectionId")

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	wayCollectionRaw, err := wc.wayCollectionFacade.UpdateWayCollection(ctx, wayCollectionID, payload.Name)
	utils.HandleErrorGin(ctx, err)

	wayCollection := schemas.WayCollectionPlainResponse{
		Uuid: wayCollectionRaw.Uuid,
		Name: wayCollectionRaw.Name,
	}

	ctx.JSON(http.StatusOK, wayCollection)
}

// Deleting way handlers
// @Summary Delete wayCollection by UUID
// @Description
// @Tags wayCollection
// @ID delete-wayCollection
// @Accept  json
// @Produce  json
// @Param wayCollectionId path string true "wayCollection ID"
// @Success 204
// @Router /wayCollections/{wayCollectionId} [delete]
func (wc *WayCollectionController) DeleteWayCollectionById(ctx *gin.Context) {
	wayCollectionID := ctx.Param("wayCollectionId")

	err := wc.wayCollectionFacade.DeleteWayCollectionById(ctx, wayCollectionID)
	utils.HandleErrorGin(ctx, err)

	ctx.Status(http.StatusNoContent)
}